requirements.txt
openai
langchain
langchain-astradb
langchain-openai
langchainhub
requests
python-dotenv
